Moved the `Serialize` and `Deserialize` implementations to a separate module
